Move temp_folder in temp_fs
